--- frontend/src/Common/Empty.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { IChallengeProps } from "../Challenge/IChallengeProps";
-
-
-const Empty = (props: IChallengeProps) => {
-  return (
-    <div>asd</div>
-  );
-};
-
-export default Empty;
