+++ frontend/src/Common/TabPanel/ITabPanelProps.ts
@@ -1,0 +1,6 @@
+export default interface ITabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  selectedTabIndex: number;
+  className: any;
+}
