+++ frontend/src/Confirmation/styles.tsx
@@ -1,0 +1,9 @@
+import { makeStyles } from "@material-ui/core";
+import { green } from "@material-ui/core/colors";
+
+const useStyles = makeStyles({
+  successButton: { color: green[500] }
+});
+
+
+export default useStyles;
