+++ frontend/src/Flag/IFlagProps.ts
@@ -1,0 +1,4 @@
+export default interface IFlagProps {
+  flag?: string;
+  resetChallenge: () => void;
+}
