+++ frontend/src/Flag/styles.tsx
@@ -1,0 +1,12 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+  root: {
+    marginBottom: '50px',
+    textAlign: 'center',
+    border: '1px solid #CCC'
+  },
+});
+
+
+export default useStyles;
